refactor import field
