+++ src/components/input-field/index.ts
@@ -1,1 +1,2 @@
 export * from "./input-field.jsx";
+export * from "./text-input-field.jsx";
